rsx Component

rsx Component

Usage

component(
  name = paste0("unnamed_component-", random_id()),
  data = NULL,
  methods = NULL,
  template = NULL,
  styles = NULL
)

Arguments

name

TODO

data

function that returns a named list

methods

named list of functions

template

function that returns a taglist

styles

named list of "global" and "scoped" styles

Details

Data

Data is used for a component's internal state and can also be used to pass information from a parent component to its children. Both component templates and methods have access to component data.

The component data function must be of the following structure:

function() {
 # must either return a named list or NULL
 list()
}

For instance, the following is a valid data function:

function() {
 list(
   rctv = shiny::reactiveVal(),
   df   = mtcars
 )
}

Methods

Methods are a list of functions that are contained within a given component, and typically will do work with component data.

The setup method is special and is treated as the equivalent of a module server for the component.

The render method is special and is called when the UI of a component is rendered.

The following is an excerpt from a simple counter component:

methods = list(
    setup = function(input, output, session) {
        output$txt <- renderText({
            paste0("Count is: ", self$count())
        })
        observeEvent(input$button, {
            self$count(self$count() + 1L)
        })
    },
    increment = function() {
        self$count(self$count() + 1L)
    }
)

Template

The template refers to the UI-generator of a given component. This is analagous to the UI function of a given shiny module.

Template requires one argument: ns. ns is used the same as in shiny modules, and is used to distinguish between component instances. The user does not need to supply their own shiny::NS call -- this is handled by rsx.

The component template function must be of the following structure:

function(ns) {
 # must either return an object that can be coerced to a `shiny::tags` object or a `shiny::tagList`
}

The following is a valid template:

function(ns) {
 shiny::div("Hello world!")
}

Styles

The styles argument is a named list of scoped and global CSS styles.

list(
 scoped = "",
 global = ""
)