Create an rsx server

rsx server is a low-level function that loads up the module servers of all instantiated rsx::components. If the rsx_app function is not viable for your shiny application setup, the rsx_server function can be used.

Usage

rsx_server()

Value

shiny server function

Details

Some things to note: * Due to randomised hashing of component namepaces, shiny modules are not nested in rsx * If a component does not have a setup function, a module will not be created for the instance

See also